Removing measuring section in Electricity.md
Once the patterns database is up and we have a section on measuring energy I think it would be best to point people to that from the principles than describe just two methods here. The suggestions here are accurate but also incomplete, there are many dozens of ways of measuring (i.e. Intel RAPL, Scaphandre, Windows E3 are just 3 but there are loads more including models which we are coming up with. I think it's better to list them all in the patterns database and point this doc to that once it's ready.